Improve performance of sinter by introducing 'sinter.CompiledDecoder' (#528)
Originally, sinter did all sampling for a batch then all decoding. It was done this way to amortize the cost of configuring the decoder at the start of the decode call. However, the batches were large enough that they thrashed memory so I switched it to storing them on disk. This worked better, but turned the disk into a bottleneck. This PR goes back to storing the shots in memory by adding a 'sinter.CompiledDecoder' shim for representing a preconfigured decoder. This allows a one-time configuration cost to be amortized over many mini batches of 1024 shots, instead of e.g. 100K shots, so the shots can be kept in memory. This change also benefits from a previous change to make stim.CompiledSampler do a better job at caching a configured frame simulator. Now sinter will call 'sinter.Decoder.compile_decoder_for_dem' to try to get one and then start passing small batches into it in order to build up the larger requested batch. In my quick testing this achieves ~2x performance on a 96 core machine. I can also tell from the progress printout and from top that it's just spending less time blocked and thrashing; utilization is much better. Also, when you interrupt it, it now closes much faster since it doesn't need to clean up as many files.