fix help_modes test
After updating run_captured_stim_main to take std::string for stdin
instead of a cstring, main.help_modes started failing. For cases where
an empty string was passed instead of nullptr, run_captured_stim_main
created a tmp file and passed it with `--in`. This caused some test
cases in help_modes to behave differently than expected.

This commit simply updates the output on a single test case to reflect
the real output of this command (`stim --help --sample`).