add `cluster_node_limit` for MWPF decoder to better tune decoding time and accuracy (#857)
This is a new parameter that is agnostic to individual machine's clock
speed. We limit the maximum number of dual variables inside each cluster
to avoid wasting computing time on small yet complicated clusters. A
default value of 50 is good enough for small code and improves the
decoding speed a lot. Intuitively (but not precisely), this means we
limit the maximum number of dual variables (blossoms and their children)
in an alternating tree to 50, and once it hits this limit, it will fall
back to union-find decoder. Of course, in hypergraph cases, the
situation is a little bit more complicated and this limit is hit more
often.