Enhance Getting Started notebook
There are no code or functional changes in this commit.

These changes were motivated by wanting to provide a version of the
notebook in Google Colab, so that the Stim landing page on
quantumai.google could link to it. Because the notebook in Colab will
be outside the context of the GitHub repository, I felt it would be
helpful to provide more contextual information at the beginning. While
at it, I saw a few minor typos to fix ... and then one thing led to
another, and more editing ensued all over.