Enhance the "Getting Started" notebook (#868)
There are no code or functional changes in this commit.

These changes were motivated by wanting to provide a version of the
notebook in Google Colab, so that the Stim landing page on
quantumai.google could link to it. Because the notebook in Colab will be
outside the context of the GitHub repository, I felt it would be helpful
to provide more contextual information at the top of the notebook. While
I was doing that, I saw a few minor typos to fix in the rest of the
document ... and then one thing led to another, and more minor editing
ensued all over. The additional edits concern some style consistency
issues, plus the addition of resources at the end.