Add config file for markdownlint
This adds a configuration file for a Markdown file
linter:(https://github.com/igorshubovych/markdownlint-cli). It follows
Google Markdown style as closely as possible in a way compatible with
GitHub-flavored Markdown, with the exception of the line length, which
is set to 120 in order to match the value in `.clang-format`.