connect max_batch_size from CollectionManager to worker RampThrottledSampler (#943)
When `max_batch_size` is passed to `sinter.collect` it gets passed to
the `CollectionOptions` for the `CollectionManager` but then in
`CollectionManager.start_workers` it is not passed on to the
`RampThrottledSampler` instead `max_batch_shots` is hardcoded to 1024
[here](https://github.com/quantumlib/Stim/blob/9156ab6c97c4f27be1160cd8b6aafb35f7677e57/glue/sample/src/sinter/_collection/_collection_manager.py#L140).
I have seen some small improvements in runtime when I can go for larger
batches.

This PR passes `max_batch_size` to
`RampThrottledSampler.max_batch_shots` and adds some test coverage.